Add Iteration class and IterationsService
Added a new Iteration class in iteration.ts and IterationsService in iterations-service.ts. Iteration class properties have been defined and methods to create, get and list iterations have been implemented in IterationsService. Also added a method in the Story class to get the Iteration instance. Updated the client.ts to include an instance of IterationsService.